pyproject.toml
[project]
name = "sparktate"
version = "0.1.0"
description = "Live speech-to-text daemon for DGX Spark using NVIDIA Parakeet"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "nemo_toolkit[asr]>=2.0.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pyperclip>=1.8.0",
    "pynput>=1.7.0",
]

[project.scripts]
sparktate = "sparktate.cli:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
